+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 # [Unreleased]
+### Fixed
+- Fixed potential panic if a benchmark took zero time.
 
 ## [1.0.0-alpha3] - 2020.07-06
 ### Added